Split tag detection (+ tests) into multiple files.
