+++ lib/src/create_report.dart
@@ -30,7 +30,7 @@ import 'model.dart';
 import 'package_context.dart';
 import 'pubspec.dart';
 import 'sdk_env.dart';
-import 'tag_detection.dart';
+import 'tag/tagger.dart';
 import 'utils.dart';
 
 const _pluginDocsUrl =
